+++ apps/faces-example1/src/main/java/org/apache/struts/webapp/example/MainMenuBacking.java
@@ -23,7 +23,9 @@
 package org.apache.struts.webapp.example;
 
 
+import javax.enterprise.context.RequestScoped;
 import javax.faces.context.FacesContext;
+import javax.inject.Named;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -33,6 +35,8 @@ import org.slf4j.LoggerFactory;
  * <p>Backing bean for the <code>mainMenu.jsp</code> page.</p>
  */
 
+@Named
+@RequestScoped
 public class MainMenuBacking extends AbstractBacking {