+++ apps/faces-example2/src/main/java/org/apache/struts/webapp/example2/LoggedOff.java
@@ -25,8 +25,10 @@ package org.apache.struts.webapp.example2;
 
 import java.io.IOException;
 
+import javax.enterprise.context.RequestScoped;
 import javax.faces.FacesException;
 import javax.faces.context.FacesContext;
+import javax.inject.Named;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -36,6 +38,8 @@ import org.slf4j.LoggerFactory;
  * <p>Backing bean for the <code>loggedoff.jsp</code> page.</p>
  */
 
+@Named
+@RequestScoped
 public class LoggedOff {